Move Styles to sub folder inside components
This makes it easy to keep track of everything.
